backend/requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.3
Flask_Cors==3.0.10
langchain==0.0.142
marqo==0.9.1
python-dotenv==1.0.0
Requests==2.28.2
llama-cpp-python
black